requirements.txt
Flask==2.3.3
pandas==2.2.2
python-dateutil==2.9.0